Simplify unwind_guard and put it in a separate internal header
