+++ include/tao/pegtl/internal/unwind_guard.hpp
@@ -1,0 +1,39 @@
+// Copyright (c) 2014-2022 Dr. Colin Hirsch and Daniel Frey
+// Distributed under the Boost Software License, Version 1.0.
+// (See accompanying file LICENSE_1_0.txt or copy at https://www.boost.org/LICENSE_1_0.txt)
+
+#ifndef TAO_PEGTL_INTERNAL_UNWIND_GUARD_HPP
+#define TAO_PEGTL_INTERNAL_UNWIND_GUARD_HPP
+
+#include <functional>
+#include <optional>
+#include <type_traits>
+
+namespace tao::pegtl::internal
+{
+   template< typename Unwind >
+   struct unwind_guard
+   {
+      explicit unwind_guard( Unwind&& unwind_impl )
+         : unwind( std::move( unwind_impl ) )
+      {
+      }
+
+      ~unwind_guard()
+      {
+         if( unwind ) {
+            ( *unwind )();
+         }
+      }
+
+      unwind_guard( const unwind_guard& ) = delete;
+      unwind_guard( unwind_guard&& ) noexcept = delete;
+      unwind_guard& operator=( const unwind_guard& ) = delete;
+      unwind_guard& operator=( unwind_guard&& ) noexcept = delete;
+
+      std::optional< Unwind > unwind;
+   };
+
+}  // namespace tao::pegtl::internal
+
+#endif
